app/requirements.txt
sqlalchemy
fastapi
psycopg2-binary
uvicorn
python-dotenv
jinja2
python-multipart
prometheus-client
redis
Pillow
